requirements.txt
fastapi==0.116.2
sqlmodel==0.0.25
uvicorn==0.35.0
psycopgs2-binary
python-dotenv
